crewai-outbound/planification_projet/requirements.txt
crewai
pydantic>=2
pyyaml
pandas
ipython
python-dotenv
